dataflow api refactor
